requirements.txt
networkx==3.4.2
numpy==2.2.4
pandas==2.2.3
psycopg2-binary
toml==0.10.2
streamlit_searchbox
fuzzywuzzy
bcrypt